feat(lidarr): add artist tagging and fix default monitoring option
      - Add getTags() method to LidarrClient and /api/settings/lidarr/tags endpoint
      - Apply configured tagId to artist on add; send tags:[] when none configured
      - Change default monitor option fallback from "none" to "all"
      - Add tag dropdown with refresh button to Settings → Integrations
      - Fix SettingsPage.jsx import paths and wire missing tag props
