+++ frontend/src/pages/Settings/hooks/useSettingsData.js
@@ -5,6 +5,7 @@ import api, {
   updateAppSettings,
   getLidarrProfiles,
   getLidarrMetadataProfiles,
+  getLidarrTags,
   testLidarrConnection,
   testGotifyConnection,
   applyLidarrCommunityGuide,
@@ -36,7 +37,8 @@ const defaultSettings = {
       apiKey: "",
       qualityProfileId: null,
       metadataProfileId: null,
-      defaultMonitorOption: "none",
+      tagId: null,
+      defaultMonitorOption: "all",
       searchOnAdd: false,
     },
     musicbrainz: { email: "" },
@@ -69,6 +71,8 @@ export function useSettingsData(showSuccess, showError, showInfo) {
   const [lidarrMetadataProfiles, setLidarrMetadataProfiles] = useState([]);
   const [loadingLidarrMetadataProfiles, setLoadingLidarrMetadataProfiles] =
     useState(false);
+  const [lidarrTags, setLidarrTags] = useState([]);
+  const [loadingLidarrTags, setLoadingLidarrTags] = useState(false);
   const [testingLidarr, setTestingLidarr] = useState(false);
   const [testingGotify, setTestingGotify] = useState(false);
   const [applyingCommunityGuide, setApplyingCommunityGuide] = useState(false);
@@ -140,17 +144,21 @@ export function useSettingsData(showSuccess, showError, showInfo) {
       if (lidarr.url && lidarr.apiKey) {
         setLoadingLidarrProfiles(true);
         setLoadingLidarrMetadataProfiles(true);
+        setLoadingLidarrTags(true);
         try {
-          const [profiles, metadataProfiles] = await Promise.all([
+          const [profiles, metadataProfiles, tags] = await Promise.all([
             getLidarrProfiles(lidarr.url, lidarr.apiKey),
             getLidarrMetadataProfiles(lidarr.url, lidarr.apiKey),
+            getLidarrTags(lidarr.url, lidarr.apiKey),
           ]);
           setLidarrProfiles(profiles);
           setLidarrMetadataProfiles(metadataProfiles);
+          setLidarrTags(Array.isArray(tags) ? tags : []);
         } catch {
         } finally {
           setLoadingLidarrProfiles(false);
           setLoadingLidarrMetadataProfiles(false);
+          setLoadingLidarrTags(false);
         }
       }
     } catch {}
@@ -359,6 +367,10 @@ export function useSettingsData(showSuccess, showError, showInfo) {
     setLidarrMetadataProfiles,
     loadingLidarrMetadataProfiles,
     setLoadingLidarrMetadataProfiles,
+    lidarrTags,
+    setLidarrTags,
+    loadingLidarrTags,
+    setLoadingLidarrTags,
     testingLidarr,
     setTestingLidarr,
     testingGotify,
